datacenters:
We're attempting to derive the proper activation keys for the various moving
parts of the ovirt system via introspection of the parameters object.  I'm not
entirely pleased with the method which comes down to ferriting out those
properties that are unique-per-object-type.  It may have more to do with my
lack of knowledge, but this seems like the only real way to do it.

Because most of ovirt-sdk is generated python (from some part of the ovirt api
code), they can afford to be verbose; however, towards the goal of creating
a more useable interface for ovirt, I have neither the inclination or time
to write individual wrappers for each outlier object-type.  On top of that,
the ovirtsdk api will undboubtedly undergo changes that would break that
kind of model.